backendAIInference/requirements.txt
firebase_admin
ultralytics
pillow